+++ pgdrive/component/vehicle_module/distance_detector.py
@@ -1,8 +1,6 @@
 import logging
-from pgdrive.engine.engine_utils import get_engine
-from collections import namedtuple
 import math
-from collections import defaultdict
+from collections import namedtuple
 
 import numpy as np
 from panda3d.bullet import BulletGhostNode, BulletSphereShape
@@ -10,120 +8,11 @@ from panda3d.core import BitMask32, NodePath
 
 from pgdrive.constants import CamMask, CollisionGroup
 from pgdrive.engine.asset_loader import AssetLoader
+from pgdrive.engine.engine_utils import get_engine
 from pgdrive.utils import import_cutils
 from pgdrive.utils.coordinates_shift import panda_position
 
 cutils = import_cutils()
-
-
-class DetectorMask:
-    def __init__(self, num_lasers: int, max_span: float, max_distance: float = 1e6):
-        raise DeprecationWarning("This class is deprecated, the new lidar is implemented with this optimization, and"
-                                 "can be easily extended to 3d")
-        self.num_lasers = num_lasers
-        self.angle_delta = 360 / self.num_lasers
-        # self.max_span = max_span
-        self.half_max_span_square = (max_span / 2)**2
-        self.masks = defaultdict(lambda: np.zeros((self.num_lasers, ), dtype=np.bool))
-        # self.max_distance = max_distance + max_span
-        self.max_distance_square = (max_distance + max_span)**2
-
-    def update_mask(self, position_dict: dict, heading_dict: dict, is_target_vehicle_dict: dict):
-        assert set(position_dict.keys()) == set(heading_dict.keys()) == set(is_target_vehicle_dict.keys())
-        if not position_dict:
-            return
-
-        for k in self.masks.keys():
-            self.masks[k].fill(False)
-
-        for k, is_target in is_target_vehicle_dict.items():
-            if is_target:
-                self.masks[k]  # Touch to create entry
-
-        keys = list(position_dict.keys())
-        for c1, k1 in enumerate(keys[:-1]):
-            for c2, k2 in enumerate(keys[c1 + 1:]):
-
-                if (not is_target_vehicle_dict[k1]) and (not is_target_vehicle_dict[k2]):
-                    continue
-
-                pos1 = position_dict[k1]
-                pos2 = position_dict[k2]
-                head1 = heading_dict[k1]
-                head2 = heading_dict[k2]
-
-                diff = (pos2[0] - pos1[0], pos2[1] - pos1[1])
-                dist_square = diff[0]**2 + diff[1]**2
-                if dist_square < self.half_max_span_square:
-                    if is_target_vehicle_dict[k1]:
-                        self._mark_all(k1)
-                    if is_target_vehicle_dict[k2]:
-                        self._mark_all(k2)
-                    continue
-
-                if dist_square > self.max_distance_square:
-                    continue
-
-                span = None
-                if is_target_vehicle_dict[k1]:
-                    span = math.asin(math.sqrt(self.half_max_span_square / dist_square))
-                    # relative heading of v2's center when compared to v1's center
-                    relative_head = math.atan2(diff[1], diff[0])
-                    head_in_1 = relative_head - head1
-                    head_in_1_max = head_in_1 + span
-                    head_in_1_min = head_in_1 - span
-                    head_1_max = np.rad2deg(head_in_1_max)
-                    head_1_min = np.rad2deg(head_in_1_min)
-                    self._mark_this_range(head_1_min, head_1_max, name=k1)
-
-                if is_target_vehicle_dict[k2]:
-                    if span is None:
-                        span = math.asin(math.sqrt(self.half_max_span_square / dist_square))
-                    diff2 = (-diff[0], -diff[1])
-                    # relative heading of v2's center when compared to v1's center
-                    relative_head2 = math.atan2(diff2[1], diff2[0])
-                    head_in_2 = relative_head2 - head2
-                    head_in_2_max = head_in_2 + span
-                    head_in_2_min = head_in_2 - span
-                    head_2_max = np.rad2deg(head_in_2_max)
-                    head_2_min = np.rad2deg(head_in_2_min)
-                    self._mark_this_range(head_2_min, head_2_max, name=k2)
-
-    def _mark_this_range(self, small_angle, large_angle, name):
-        # We use clockwise to determine small and large angle.
-        # For example, if you wish to fill 355 deg to 5 deg, then small_angle is 355, large_angle is 5.
-        small_angle = small_angle % 360
-        large_angle = large_angle % 360
-
-        assert 0 <= small_angle <= 360
-        assert 0 <= large_angle <= 360
-
-        small_index = math.floor(small_angle / self.angle_delta)
-        large_index = math.ceil(large_angle / self.angle_delta)
-        if large_angle < small_angle:  # We are in the case like small=355, large=5
-            self.masks[name][small_index:] = True
-            self.masks[name][:large_index + 1] = True
-        else:
-            self.masks[name][small_index:large_index + 1] = True
-
-    def _mark_all(self, name):
-        self.masks[name].fill(True)
-
-    def get_mask(self, name):
-        assert name in self.masks, "It seems that you have not initialized the mask for vehicle {} yet!".format(name)
-        return self.masks[name]
-
-    def clear(self):
-        self.masks.clear()
-
-    def get_mask_ratio(self):
-        total = 0
-        masked = 0
-        for k, v in self.masks.items():
-            total += v.size
-            masked += v.sum()
-        return masked / total
-
 
 detect_result = namedtuple("detect_result", "cloud_points detected_objects")
 
@@ -152,7 +41,6 @@ class DistanceDetector:
         self.start_phase_offset = 0
         self.origin = parent_node_np.attachNewNode("Could_points")
         self._lidar_range = np.arange(0, self.num_lasers) * self.radian_unit + self.start_phase_offset
-        self.extra_filter_node= {parent_node_np.node()}
 
         # override these properties to decide which elements to detect and show
         self.origin.hide(CamMask.RgbCam | CamMask.Shadow | CamMask.Shadow | CamMask.DepthCam)
@@ -174,7 +62,7 @@ class DistanceDetector:
             # self.origin.flattenStrong()
 
     def perceive(self, base_vehicle, physics_world, detector_mask: np.ndarray = None):
-        extra_filter_node = self.extra_filter_node
+        extra_filter_node = {base_vehicle.origin.node()}
         vehicle_position = base_vehicle.position
         heading_theta = base_vehicle.heading_theta
         assert not isinstance(detector_mask, str), "Please specify detector_mask either with None or a numpy array."
