+++ pgdrive/manager/replay_record_system.py
@@ -47,7 +47,7 @@ class Replayer:
                     vehicle_to_set.set_state(t_v_s)
                     if vehicle_to_set.routing_localization.final_road != Road(*t_v_s["destination"]):
                         vehicle_to_set.routing_localization.set_route(t_v_s["spawn_road"][0], t_v_s["destination"][-1])
-                    vehicle_to_set.after_step(detector_mask=None)
+                    vehicle_to_set.after_step()
             elif index == TRAFFIC_VEHICLES:
                 for t_v_idx, t_v_s in state.items():
                     vehicle_to_set = self.restore_vehicles[t_v_idx]
